pyproject.toml
[project]
name = "mcp-intro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "cryptography>=45.0.3",
    "duckdb>=1.2.2",
    "google-ai-generativelanguage>=0.6.1",
    "google-genai>=1.19.0",
    "httpx>=0.28.1",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "langchain-mcp-adapters>=0.0.11",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.3",
    "mcp-server-git>=2025.1.14",
    "mcp[cli]>=1.8.0",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "uv>=0.7.3",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]

[tool.uv.workspace]
members = [
    "scout/TestDir/Project_Alpha",
    "scout/TestDir/my_project",
]
